# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Disciplr API
  version: 0.1.0
  description: API documentation for Disciplr backend
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ExportJobResponse:
      type: object
      properties:
        jobId:
          type: string
          description: Export job identifier
          example: job_abc123
        statusUrl:
          type: string
          description: URL to poll for job status
          example: /api/exports/status/job_abc123
        pollIntervalMs:
          type: number
          description: Recommended polling interval in milliseconds
          example: 1000
      required:
        - jobId
        - statusUrl
        - pollIntervalMs
    ExportJobStatus:
      type: object
      properties:
        jobId:
          type: string
          example: job_abc123
        status:
          type: string
          enum:
            - pending
            - running
            - done
            - failed
          example: done
        attempts:
          type: number
          example: 1
        maxAttempts:
          type: number
          example: 3
        completedAt:
          type: string
          format: date-time
          example: 2026-06-02T10:00:00Z
        downloadUrl:
          type: string
          description: Download URL (present when status is done)
          example: /api/exports/download/token_xyz
        expiresInSeconds:
          type: number
          example: 3600
        error:
          type: string
          description: Error message if status is failed
      required:
        - jobId
        - status
        - attempts
        - maxAttempts
...
paths:
...
  /api/exports/me:
    post:
      summary: Enqueue a personal data export job
      tags:
        - Exports
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            enum:
              - csv
              - json
            description: Output format for the export
            example: json
          required: true
          description: Output format for the export
          name: format
          in: query
        - schema:
            type: string
            enum:
              - vaults
              - transactions
              - analytics
              - all
            description: Data scope for the export
            example: all
          required: true
          description: Data scope for the export
          name: scope
          in: query
      responses:
        "202":
          description: Export job enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportJobResponse"
        "400":
          description: Invalid format or scope
          content:
            application/json:
              schema: {}
        "409":
          description: Idempotency key conflict
        "429":
          description: Export quota exceeded
  /api/exports/admin:
    post:
      summary: Enqueue an admin export job (all users or target user)
      tags:
        - Exports
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            enum:
              - csv
              - json
            description: Output format for the export
            example: json
          required: true
          description: Output format for the export
          name: format
          in: query
        - schema:
            type: string
            enum:
              - vaults
              - transactions
              - analytics
              - all
            description: Data scope for the export
            example: all
          required: true
          description: Data scope for the export
          name: scope
          in: query
        - schema:
            type: string
            description: Target user ID (admin-only exports)
            example: user_123
          required: false
          description: Target user ID (admin-only exports)
          name: targetUserId
          in: query
      responses:
        "202":
          description: Export job enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportJobResponse"
        "400":
          description: Invalid format or scope
          content:
            application/json:
              schema: {}
        "403":
          description: Admin access required
        "409":
          description: Idempotency key conflict
        "429":
          description: Export quota exceeded
  /api/exports/status/{jobId}:
    get:
      summary: Poll export job status
      tags:
        - Exports
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Export job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportJobStatus"
        "403":
          description: Access denied
        "404":
          description: Job not found
  /api/exports/download/{token}:
    get:
      summary: Download completed export file
      tags:
        - Exports
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Signed download token from status response
      responses:
        "200":
          description: Export file content
          content:
            application/json:
              schema: {}
            text/csv:
              schema: {}
        "401":
          description: Invalid or expired download token
        "404":
          description: Export not ready or not found
...